Refactor: 작성한 코드 리팩토링
