+++ src/api/fetchMyOrder.ts
@@ -1,9 +1,7 @@
-const fetchMyOrder = (assetOrder: any): Promise<object> => {
+export const fetchMyOrder = (assetOrder: any): Promise<object> => {
   return new Promise((resolve) => {
     setTimeout(() => {
       resolve(assetOrder);
     }, 1000);
   });
 };
-
-export { fetchMyOrder };
